解决互斥信号量卡线程问题

+++ 工程 - 云台/FUNTION/ramp.c
@@ -8,7 +8,7 @@ void ramp_init(ramp_t *ramp,uint32_t set_cnt)
 }
 
 
-int ramp_calc(ramp_t *ramp)
+float ramp_calc(ramp_t *ramp)
 {
   if(ramp->set_count <= 0)
     return 0;
